CLIM-1447: Atom20 getPeriods taking DECADAL into account
