+++ apps/src/lib/s2d/index.ts
@@ -1,3 +1,4 @@
+export * from './calculate-periods';
 export * from './forecast-probabilities-bar-chart';
 export * from './forecast-probabilities-categories';
 export * from './utils';
